refactor: move state management
