+++ src/app/core/store/general/general-store.module.ts
@@ -3,8 +3,6 @@ import { EffectsModule } from '@ngrx/effects';
 import { ActionReducerMap, StoreModule } from '@ngrx/store';
 import { pick } from 'lodash-es';
 
-import { ContactEffects } from './contact/contact.effects';
-import { contactReducer } from './contact/contact.reducer';
 import { CountriesEffects } from './countries/countries.effects';
 import { countriesReducer } from './countries/countries.reducer';
 import { GeneralState } from './general-store';
@@ -14,10 +12,9 @@ import { regionsReducer } from './regions/regions.reducer';
 const generalReducers: ActionReducerMap<GeneralState> = {
   countries: countriesReducer,
   regions: regionsReducer,
-  contact: contactReducer,
 };
 
-const generalEffects = [ContactEffects, CountriesEffects, RegionsEffects];
+const generalEffects = [CountriesEffects, RegionsEffects];
 
 @NgModule({
   imports: [EffectsModule.forFeature(generalEffects), StoreModule.forFeature('general', generalReducers)],
