+++ src/app/extensions/contact-us/facades/contact-us.facade.ts
@@ -2,13 +2,8 @@ import { Injectable } from '@angular/core';
 import { Store, select } from '@ngrx/store';
 
 import { Contact } from 'ish-core/models/contact/contact.model';
-import {
-  createContact,
-  getContactLoading,
-  getContactSubjects,
-  getContactSuccess,
-  loadContact,
-} from 'ish-core/store/general/contact';
+
+import { createContact, getContactLoading, getContactSubjects, getContactSuccess, loadContact } from '../store/contact';
 
 /* eslint-disable @typescript-eslint/member-ordering */
 @Injectable({ providedIn: 'root' })
