+++ src/app/extensions/contact-us/store/contact-us-store.ts
@@ -1,5 +1,9 @@
 import { createFeatureSelector } from '@ngrx/store';
 
-export interface ContactUsState {}
+import { ContactState } from './contact/contact.reducer';
+
+export interface ContactUsState {
+  contact: ContactState;
+}
 
 export const getContactUsState = createFeatureSelector<ContactUsState>('contactUs');
