+++ src/app/extensions/contact-us/store/contact/contact.selectors.spec.ts
@@ -2,9 +2,10 @@ import { TestBed } from '@angular/core/testing';
 
 import { Contact } from 'ish-core/models/contact/contact.model';
 import { CoreStoreModule } from 'ish-core/store/core/core-store.module';
-import { GeneralStoreModule } from 'ish-core/store/general/general-store.module';
 import { makeHttpError } from 'ish-core/utils/dev/api-service-utils';
 import { StoreWithSnapshots, provideStoreSnapshots } from 'ish-core/utils/dev/ngrx-testing';
+
+import { ContactUsStoreModule } from '../contact-us-store.module';
 
 import {
   createContact,
@@ -31,7 +32,7 @@ describe('Contact Selectors', () => {
 
   beforeEach(() => {
     TestBed.configureTestingModule({
-      imports: [CoreStoreModule.forTesting(), GeneralStoreModule.forTesting('contact')],
+      imports: [ContactUsStoreModule.forTesting('contact'), CoreStoreModule.forTesting()],
       providers: [provideStoreSnapshots()],
     });
